+++ Changelog.md
@@ -3,6 +3,10 @@
 
 ## [UnReleased v0.7]
 
+## 2021-10-28/ml
+### added
+- Navigation Service added
+- NavBack Directive added (fs-nav-back)
 
 ## 2021-10-04/ml
 ### added